BskyRepostBot/requirements.txt
playwright
atproto
pillow
python-dotenv
feedparser
requests
flask